Update: play modes to single play file Remove: Unecessary Files Code Clean Up and Comments
