+++ codebase/config.lua
@@ -1,17 +1,17 @@
 application = {
-    content = {
-        width = 1920,
-        height = 1080,
-        scale = "letterbox",
-        xAlign = "center",
-        yAlign = "center",
-        fps = 60,
-                imageSuffix =
-                {
+   content = {
+      width = 1920,
+      height = 1080,
+      scale = "letterbox",
+      xAlign = "center",
+      yAlign = "center",
+      fps = 60,
+      imageSuffix =
+      {
 
-                   ["@2x"] = 1.5, -- for iPhone, iPod touch, iPad1, and iPad2
-                   ["@4x"] = 3.0, -- for iPad 3
+         ["@2x"] = 1.5, -- for iPhone, iPod touch, iPad1, and iPad2
+         ["@4x"] = 3.0, -- for iPad 3
 
-                },
-    }
+      },
+   }
 }
