+++ codebase/globalData.lua
@@ -5,15 +5,22 @@
 -- These variables can be altered for the current gaming session and will return to their hard-coded values when the game is restarted.
 
 local globalData = {
-    gameTitle               = "Learn to Type",
-    gameVersion             = "0.1.0",
-    gameDetails             = {},
-    isGameRunning           = false,
-    isGameOver              = false,
-    sessionDetails          = {scene={},score="0"},
-    w                       = display.actualContentWidth,
-    h                       = display.actualContentHeight,
-    timeUnit                = 500,
+   gameTitle               = "Learn to Type",
+   gameVersion             = "0.1.0",
+   gameDetails             = {},
+   isGameRunning           = false,
+   isGameOver              = false,
+   sessionDetails          = {
+                              currentScene={},
+                              score="0",
+                              mode="Word",
+                              mainLetter={},
+                              mainWord={},
+                              mainCountDown={}
+                           },
+   w                       = display.actualContentWidth,
+   h                       = display.actualContentHeight,
+   timeUnit                = 500,
 }
 
 return globalData
